fix (#21) : API, File 에 대한 Command 및 Query 처리 Service 이동
