+++ api/src/main/java/com/yapp/api/domain/file/persistence/handler/FileCommandHandlerImpl.java
@@ -1,13 +1,9 @@
 package com.yapp.api.domain.file.persistence.handler;
 
-import java.util.List;
-import java.util.Optional;
 import java.util.function.Consumer;
-import java.util.function.Function;
 
 import org.springframework.stereotype.Component;
 
-import com.yapp.api.domain.file.persistence.entity.File;
 import com.yapp.api.domain.file.persistence.repository.FileRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -20,16 +16,6 @@ public class FileCommandHandlerImpl implements FileCommandHandler {
 	@Override
 	public void save(Consumer<FileRepository> consumer) {
 		consumer.accept(fileRepository);
-	}
-
-	@Override
-	public List<File> findList(Function<FileRepository, List<File>> function) {
-		return function.apply(fileRepository);
-	}
-
-	@Override
-	public Optional<File> findOne(Function<FileRepository, Optional<File>> function) {
-		return function.apply(fileRepository);
 	}
 
 	@Override
