+++ api/src/main/java/com/yapp/api/domain/file/persistence/handler/FileQueryHandler.java
@@ -1,7 +1,7 @@
 package com.yapp.api.domain.file.persistence.handler;
 
+import java.util.List;
 import java.util.Optional;
-import java.util.function.Consumer;
 import java.util.function.Function;
 
 import com.yapp.api.domain.file.persistence.entity.File;
@@ -9,4 +9,6 @@ import com.yapp.api.domain.file.persistence.repository.FileRepository;
 
 public interface FileQueryHandler {
 	Optional<File> findOne(Function<FileRepository, Optional<File>> function);
+
+	List<File> findList(Function<FileRepository, List<File>> function);
 }
